Code used for exps in Sol

+++ bo/utils/volume.py
@@ -1,9 +1,12 @@
 
+import numpy as np 
 
 def compute_volume(region):
     dimensions = len(region)
     volume = 1
+    # print("dimensions: ",dimensions)
     for i in range(dimensions):
-        volume *= region[i][1] - region[i][0]
+        # print('volume: ',volume)
+        volume *= np.float128(region[i][1] - region[i][0])
     return volume
